Initial draft of option pricing code
Basic objects I want to be a part of the package are, for now, just kept in a single jupyter notebook. The notebook is good enough to solve toy textbook problems on European-style options, ie. from Ch.1 of Steve Shreve's stochastic calculus for finance book.  Later on a lot of the "hard" code will be placed under the hood, and will input an easier UI.